refactor: share Card component styles
