+++ src/component/Card.component.tsx
@@ -1,51 +1,9 @@
 import React from 'react';
-import { StyleProp, StyleSheet, Text, TextStyle, TouchableOpacity, View } from 'react-native';
+import { StyleProp, Text, TextStyle, TouchableOpacity, View } from 'react-native';
 
-import { DEFAULT_DARK_THEME, DEFAULT_LIGHT_THEME, Theme, useTheme, useThemeAwareObject } from '../theme';
+import { DEFAULT_DARK_THEME, DEFAULT_LIGHT_THEME, useTheme, useThemeAwareObject } from '../theme';
+import { createStyles } from './Card.styles';
 import { RenderCounter } from './RenderCounter.component';
-
-const createStyles = (theme: Theme) => {
-  const styles = StyleSheet.create({
-    root: {
-      alignItems: 'center',
-      backgroundColor: theme.color.surface,
-      justifyContent: 'center',
-      marginTop: theme.spacing.double,
-      padding: theme.spacing.double,
-    },
-    infoTextWrapper: {
-      flexDirection: 'row',
-      marginBottom: theme.spacing.double,
-    },
-    infoText: {
-      color: theme.color.onSurface,
-      fontSize: 14,
-    },
-    infoTextBold: {
-      fontWeight: 'bold',
-    },
-    buttonsWrapper: {
-      flexDirection: 'row',
-      justifyContent: 'space-between',
-      marginTop: theme.spacing.double,
-    },
-    button: {
-      backgroundColor: theme.color.primary,
-      borderRadius: 4,
-      flex: 1,
-      justifyContent: 'center',
-      marginHorizontal: theme.spacing.base,
-      padding: theme.spacing.base,
-    },
-    buttonText: {
-      color: theme.color.onPrimary,
-      fontSize: 14,
-      fontWeight: 'bold',
-      textAlign: 'center',
-    },
-  });
-  return styles;
-};
 
 export const Card = React.memo(() => {
   const { theme, setTheme, toggleTheme } = useTheme();
